Introduce U128<Balance> in xpallet_support
