+++ xpallets/mining/staking/rpc/Cargo.toml
@@ -17,3 +17,4 @@ jsonrpc-derive = "14.2.1"
 jsonrpc-core-client = "14.2.0"
 xpallet-mining-staking = { path = "../" }
 xpallet-mining-staking-rpc-runtime-api = { path = "./runtime-api" }
+xpallet-support = { path = "../../../support" }
