+++ xpallets/mining/staking/rpc/runtime-api/Cargo.toml
@@ -11,9 +11,9 @@ targets = ["x86_64-unknown-linux-gnu"]
 codec = { package = "parity-scale-codec", version = "1.3.1", default-features = false }
 sp-api = { git = "https://github.com/paritytech/substrate.git", tag = "v2.0.0-rc4", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate.git", tag = "v2.0.0-rc4", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git", tag = "v2.0.0-rc4", default-features = false }
 
 xpallet-mining-staking = { path = "../..", default-features = false }
+xpallet-support = { path = "../../../../support", default-features = false }
 
 [features]
 default = ["std"]
@@ -21,7 +21,7 @@ std = [
     "codec/std",
     "sp-api/std",
     "sp-std/std",
-    "sp-runtime/std",
 
     "xpallet-mining-staking/std",
+    "xpallet-support/std",
 ]
