+++ xpallets/support/Cargo.toml
@@ -5,7 +5,9 @@ authors = ["ChainX community <https://www.chainx.org>"]
 edition = "2018"
 
 [dependencies]
+codec = { package = "parity-scale-codec", version = "1.3.1", features = ["derive"], default-features = false }
 rustc-hex = { version = "2.0", optional = true }
+serde = { version = "1.0", optional = true }
 
 # Substrate primitives
 sp-std = { git = "https://github.com/paritytech/substrate.git", tag = "v2.0.0-rc4", default-features = false }
@@ -16,7 +18,9 @@ frame-support = { git = "https://github.com/paritytech/substrate.git", tag = "v2
 [features]
 default = ["std"]
 std = [
+    "codec/std",
     "rustc-hex",
+    "serde",
 
     "sp-std/std",
